requirements.txt
python-dotenv==1.0.1
litellm==1.83.4
python-magic-bin==0.4.14
pdfplumber==0.11.9
PyMuPDF==1.27.2.2
python-docx==1.2.0
trafilatura==2.0.0
pandas==3.0.2
openpyxl==3.1.5
requests==2.33.1
